iis.py tries to load on linux with errors
from https://docs.saltstack.com/en/latest/ref/modules/#virtual-function:
"""
Since __virtual__ is called before the module is loaded, 
__salt__ will be unavailable as it will not have been packed into the module at this point in time.
"""

On linux this state gets loaded, with an error because of the above reason. I understand that you want to only load the state module if the pillar module is also loaded, but the current method only works by coincidence.  The IIS pillar module uses salt.utils.is_windows(), so I thought that would be the best way to do this.